+++ backend/zog_igame/models/__init__.py
@@ -1,14 +1,19 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import deal
 from . import game
 from . import team
-from . import deal
 from . import match
 from . import table
 from . import board
-from . import board1_bid
-from . import board2_play
 from . import board_call
 from . import board_card
+
+from . import table_board
+from . import match_table
+from . import team_score
+
+from . import board1_bid
+from . import board2_play
 from . import channel
